Move ssh client functions from utils to services
