+++ src/dstack/_internal/cli/main.py
@@ -19,8 +19,8 @@ from dstack._internal.cli.commands.volume import VolumeCommand
 from dstack._internal.cli.utils.common import _colors, console
 from dstack._internal.cli.utils.updates import check_for_updates
 from dstack._internal.core.errors import ClientError, CLIError, ConfigurationError, SSHError
+from dstack._internal.core.services.ssh.client import get_ssh_client_info
 from dstack._internal.utils.logging import get_logger
-from dstack._internal.utils.ssh import get_ssh_client_info
 from dstack.version import __version__ as version
 
 logger = get_logger(__name__)
